Python from Julia

Execute Python code

PyObjects

Wrapper types

Python.PyList — Type

PyList{T=PyObject}([o])

Wrap the Python list o (or anything satisfying the sequence interface) as a Julia vector with elements of type T.

If o is not given, an empty list is created.

Python.PySet — Type

PySet{T=PyObject}([o])

Wrap the Python set o (or anything satisfying the set interface) as a Julia set with elements of type T.

If o is not given, an empty set is created.

Python.PyDict — Type

PyDict{K=PyObject, V=PyObject}([o])

Wrap the Python dictionary o (or anything satisfying the mapping interface) as a Julia dictionary with keys of type K and values of type V.

If o is not given, an empty dict is created.

Python.PyIterable — Type

PyIterable{T=PyObject}(o)

Wrap the Python object o into a Julia object which iterates values of type T.

Python.PyArray — Type

PyArray{T,N,R,M,L}(o)

Interpret the Python array o as a Julia array.

The input may be anything supporting the buffer protocol or the numpy array interface. This includes bytes, bytearray, array.array, numpy.ndarray, pandas.Series.

All type parameters are optional:

T is the (Julia) element type.

N is the number of dimensions.

R is the type of elements of the underlying buffer (which may be different from T to allow some basic conversion).

M is true if the array is mutable.

L is true if the array supports fast linear indexing.

Python.PyBuffer — Type

PyBuffer(o, [flags=C.PyBUF_FULL_RO])

A reference to the underlying buffer of o, if it satisfies the buffer protocol.

Has the following properties:

buf: Pointer to the data.

obj: The exporting object (usually o).

len: The length of the buffer in bytes.

readonly: True if the buffer is immutable.

itemsize: The size of each element.

format: The struct-syntax format of the element type.

ndim: The number of dimensions.

shape: The length of the buffer in each dimension.

strides: The strides (in bytes) of the buffer in each dimension.

suboffsets: For indirect arrays. See the buffer protocol documentation.

isccontiguous: True if the buffer is C-contiguous (e.g. numpy arrays).

isfcontiguous: True if the buffer is Fortran-contiguous (e.g. Julia arrays).

eltype: The element type.

Python.PyIO — Type

PyIO(o; own=false, text=missing, buflen=4096)

Wrap the Python byte-based IO stream o as a Julia IO stream.

When this goes out of scope and is finalized, it is automatically flushed. If own=true then it is also closed.

If text=false then o must be a binary stream and arbitrary binary I/O is possible. If text=true then o must be a text stream and only UTF-8 must be written (i.e. use print not write). If text is not specified then it is chosen automatically. If o is a text stream and you really need a binary stream, then often PyIO(o.buffer) will work.

For efficiency, reads and writes are buffered before being sent to o. The size of the buffer is buflen.

Python.PyPandasDataFrame — Type

PyPandasDataFrame(o; indexname=:index, columntypes=(), copy=false)

Wrap the Pandas dataframe o as a Julia table.

This object satisfies the Tables.jl and TableTraits.jl interfaces.

:indexname is the name of the index column when converting this to a table, and may be nothing to exclude the index.

:columntypes is an iterable of columnname=>type or [columnnames...]=>type pairs, used when converting to a table.

:copy is true to copy columns on conversion.

Conversion to Python

Conversion from Python