Julia from Python

Wrapper types

Apart from a few fundamental immutable types (see here), all Julia values are by default converted into Python to some julia.AnyValue object, which wraps the original value. Some types are converted to a subclass of julia.AnyValue which provides additional Python semantics –- e.g. Julia vectors are interpreted as Python sequences.

There is also a julia.RawValue object, which gives a stricter "Julia-only" interface, documented below. These types all inherit from julia.ValueBase:

julia.ValueBase

julia.RawValue

julia.AnyValue

julia.NumberValue

julia.ComplexValue

julia.RealValue

julia.RationalValue

julia.IntegerValue

julia.ArrayValue

julia.VectorValue

julia.DictValue

julia.SetValue

julia.IOValue

julia.RawIOValue

julia.BufferedIOValue

julia.TextIOValue

julia.ModuleValue

julia.TypeValue

julia.AnyValue — Python Class

Members

__jl_raw(): Convert to a julia.RawValue

julia.NumberValue — Python Class

This wraps any Julia Number value. It is a subclass of numbers.Number and behaves similar to other Python numbers.

There are also subtypes julia.ComplexValue, julia.RealValue, julia.RationalValue, julia.IntegerValue which wrap values of the corresponding Julia types, and are subclasses of the corresponding numbers ABC.

julia.ArrayValue — Python Class

This wraps any Julia AbstractArray value. It is a subclass of collections.abc.Collection.

It supports zero-up indexing, and can be indexed with integers or slices. Slicing returns a view of the original array.

There is also the subtype julia.VectorValue which wraps any AbstractVector. It is a subclass of collections.abc.Sequence and behaves similar to a Python list.

If the array is strided and its eltype is supported (i.e. Bool, IntXX, UIntXX, FloatXX, Complex{FloatXX}, Ptr{Cvoid} or Tuple or NamedTuple of these) then it supports the buffer protocol and the numpy array interface. This means that numpy.asarray(this) will yield a view of the original array, so mutations are visible on the original.

Otherwise, the numpy __array__ method is supported, and this returns an array of Python objects converted from the contents of the array. In this case, numpy.asarray(this) is a copy of the original array.

Members

ndim: The number of dimensions.

shape: Tuple of lengths in each dimension.

copy(): A copy of the array.

reshape(shape): A reshaped view of the array.

julia.DictValue — Python Class

This wraps any Julia AbstractDict value. It is a subclass of collections.abc.Mapping and behaves similar to a Python dict.

julia.SetValue — Python Class

This wraps any Julia AbstractSet value. It is a subclass of collections.abc.Set and behaves similar to a Python set.

julia.IOValue — Python Class

This wraps any Julia IO value. It is a subclass of io.IOBase.

There are also subtypes julia.RawIOValue, julia.BufferedIOValue and julia.TextIOValue, which are subclasses of io.RawIOBase (unbuffered bytes), io.BufferedIOBase (buffered bytes) and io.TextIOBase (text).

Members

torawio(): Convert to a julia.RawIOValue, an un-buffered bytes file-like object. (See also pyrawio.)
tobufferedio(): Convert to a julia.BufferedIOValue, an buffered bytes file-like object. Julia IO objects are converted to this by default. (See also pybufferedio.)
totextio(): Convert to a julia.TextIOValue, a text file-like object. (See also pytextio.)

julia.ModuleValue — Python Class

This wraps any Julia Module value.

It is the same as julia.AnyValue except for one additional convenience method:

seval([module=self], code): Evaluates the given code (a string) in the given module.

julia.TypeValue — Python Class

This wraps any Julia Type value.

It is the same as julia.AnyValue except that indexing is used to access Julia's "curly" syntax for specifying parametric types:

from julia import Main as jl
jl.Vector[jl.Int]() # equivalent to Vector{Int}() in Julia

julia.RawValue — Python Class

This can wrap any Julia value.

Members

__jl_any(): Convert to a julia.AnyValue (or subclass). (See also pyjl.)