The Julia module PythonCall

Py

PythonCall.Py — Type

Py(x)

Convert x to a Python object.

The object pybuiltins has all the standard Python builtin objects as its properties. Hence you can access pybuiltins.None and pybuiltins.TypeError.

@py

PythonCall.@py — Macro

@py expr

Evaluate the given expression using Pythonic semantics.

For example:

f(x, y) is translated to pycall(f, x, y)

x + y is translated to pyadd(x, y)

x === y is translated to pyis(x, y)

x.foo is translated to pygetattr(x, "foo")

Compound statements such as begin, if, while and for are supported.

See the online documentation for more details.

Python functions

Most of the functions in this section are essentially Python builtins with a py prefix. For example pyint(x) converts x to a Python int and is equivalent to int(x) in Python when x is a Python object.

Notable exceptions are:

pyconvert to convert a Python object to a Julia object.

pyimport to import a Python module.

pyjl to directly wrap a Julia object as a Python object.

pyclass to construct a new class.

pywith to emulate the Python with statement.

Construct Python objects

These functions convert Julia values into Python objects of standard types.

PythonCall.pybool — Function

pybool(x)

Convert x to a Python bool.

PythonCall.pyint — Function

pyint(x=0)

Convert x to a Python int.

PythonCall.pyfloat — Function

pyfloat(x=0.0)

Convert x to a Python float.

PythonCall.pycomplex — Function

pycomplex(x=0.0)
pycomplex(re, im)

Convert x to a Python complex, or create one from given real and imaginary parts.

PythonCall.pystr — Function

pystr(x)

Convert x to a Python str.

PythonCall.pybytes — Function

pybytes(x)

Convert x to a Python bytes.

PythonCall.pytuple — Function

pytuple(x=())

Convert x to a Python tuple.

If x is a Python object, this is equivalent to tuple(x) in Python. Otherwise x must be iterable.

PythonCall.pylist — Function

pylist(x=())

Convert x to a Python list.

If x is a Python object, this is equivalent to list(x) in Python. Otherwise x must be iterable.

PythonCall.pycollist — Function

pycollist(x::AbstractArray)

Create a nested Python list-of-lists from the elements of x. For matrices, this is a list of columns.

PythonCall.pyrowlist — Function

pyrowlist(x::AbstractArray)

Create a nested Python list-of-lists from the elements of x. For matrices, this is a list of rows.

PythonCall.pyset — Function

pyset(x=())

Convert x to a Python set.

If x is a Python object, this is equivalent to set(x) in Python. Otherwise x must be iterable.

PythonCall.pyfrozenset — Function

pyfrozenset(x=())

Convert x to a Python frozenset.

If x is a Python object, this is equivalent to frozenset(x) in Python. Otherwise x must be iterable.

PythonCall.pydict — Function

pydict(x)
pydict(; x...)

Convert x to a Python dict. In the second form, the keys are strings.

If x is a Python object, this is equivalent to dict(x) in Python. Otherwise x must iterate over key-value pairs.

PythonCall.pyslice — Function

pyslice([[start], stop], [step])

Construct a Python slice. Unspecified arguments default to None.

PythonCall.pyrange — Function

pyrange([[start], [stop]], [step])

Construct a Python range. Unspecified arguments default to None.

PythonCall.pymethod — Function

pymethod(x)

Convert callable x to a Python instance method.

PythonCall.pytype — Function

pytype(x)

The Python type of x.

PythonCall.pyclass — Function

pyclass(name, bases=(); members...)

Construct a new Python type with the given name, bases and members.

Equivalent to pytype(name, bases, members).

Wrap Julia values

These functions wrap Julia values into Python objects, documented here.

PythonCall.pyjl — Function

pyjl([t], x)

Create a Python object wrapping the Julia object x.

If x is mutable, then mutating the returned object also mutates x, and vice versa.

Its Python type is normally inferred from the type of x, but can be specified with t.

For example if x is an AbstractVector then the object will have type juliacall.VectorValue. This object will satisfy the Python sequence interface, so for example uses 0-up indexing.

PythonCall.pyjlraw — Function

pyjlraw(v)

Create a Python object wrapping the Julia object x.

It has type juliacall.RawValue. This has a much more rigid "Julian" interface than pyjl(v). For example, accessing attributes or calling this object will always return a RawValue.

PythonCall.pyisjl — Function

pyisjl(x)

Test whether x is a wrapped Julia value, namely an instance of juliacall.ValueBase.

PythonCall.pyjlvalue — Function

pyjlvalue(x)

Extract the value from the wrapped Julia value x.

PythonCall.pytextio — Function

pytextio(io::IO)

Wrap io as a Python text IO object.

PythonCall.pybinaryio — Function

pybinaryio(io::IO)

Wrap io as a Python binary IO object.

This is the default behaviour of Py(io).

Python builtins

PythonCall.pyconvert — Function

pyconvert(T, x, [d])

Convert the Python object x to a T.

If d is specified, it is returned on failure instead of throwing an error.

PythonCall.@pyconvert — Macro

@pyconvert(T, x, [onfail])

Convert the Python object x to a T.

On failure, evaluates to onfail, which defaults to return pyconvert_unconverted() (mainly useful for writing conversion rules).

PythonCall.pyimport — Function

pyimport(m)
pyimport(m => k)
pyimport(m => (k1, k2, ...))
pyimport(m1, m2, ...)

Import a module m, or an attribute k, or a tuple of attributes.

If several arguments are given, return the results of importing each one in a tuple.

PythonCall.pyimport_conda — Function

pyimport_conda(m, package; [channel])

Equivalent to pyimport(m) except that on failure we install the given package using conda and try again.

PythonCall.pywith — Function

pywith(f, o, d=nothing)

Equivalent to with o as x: f(x) in Python, where x is a PyObject.

On success, the value of f(x) is returned.

If an exception occurs but is suppressed then d is returned.

PythonCall.pyis — Function

pyis(x, y)

True if x and y are the same Python object. Equivalent to x is y in Python.

PythonCall.pyrepr — Function

pyrepr(x)

Equivalent to repr(x) in Python.

PythonCall.pyascii — Function

pyascii(x)

Equivalent to ascii(x) in Python.

PythonCall.pyhasattr — Function

pyhasattr(x, k)

Equivalent to hasattr(x, k) in Python.

PythonCall.pygetattr — Function

pygetattr(x, k)

Equivalent to getattr(x, k) or x.k in Python.

PythonCall.pysetattr — Function

pysetattr(x, k, v)

Equivalent to setattr(x, k, v) or x.k = v in Python.

PythonCall.pydelattr — Function

pydelattr(x, k)

Equivalent to delattr(x, k) or del x.k in Python.

PythonCall.pydir — Function

pydir(x)

Equivalent to dir(x) in Python.

PythonCall.pycall — Function

pycall(f, args...; kwargs...)

Call the Python object f with the given arguments.

PythonCall.pylen — Function

pylen(x)

The length of x. Equivalent to len(x) in Python, converted to an Integer.

PythonCall.pycontains — Function

pycontains(x, v)

Equivalent to v in x in Python.

PythonCall.pyin — Function

pyin(v, x)

Equivalent to v in x in Python.

PythonCall.pygetitem — Function

pygetitem(x, k)

Equivalent to getitem(x, k) or x[k] in Python.

PythonCall.pysetitem — Function

pysetitem(x, k, v)

Equivalent to setitem(x, k, v) or x[k] = v in Python.

PythonCall.pydelitem — Function

pydelitem(x, k)

Equivalent to delitem(x, k) or del x[k] in Python.

PythonCall.pytruth — Function

pytruth(x)

The truthyness of x. Equivalent to bool(x) in Python, converted to a Bool.

PythonCall.pyissubclass — Function

pyissubclass(s, t)

Test if s is a subclass of t. Equivalent to issubclass(s, t) in Python.

PythonCall.pyisinstance — Function

pyisinstance(x, t)

Test if x is of type t. Equivalent to isinstance(x, t) in Python.

PythonCall.pyhash — Function

pyhash(x)

Equivalent to hash(x) in Python, converted to an Integer.

PythonCall.pyiter — Function

pyiter(x)

Equivalent to iter(x) in Python.

Numeric functions

PythonCall.pyneg — Function

pyneg(x)

Equivalent to -x in Python.

PythonCall.pypos — Function

pypos(x)

Equivalent to +x in Python.

PythonCall.pyabs — Function

pyabs(x)

Equivalent to abs(x) in Python.

PythonCall.pyinv — Function

pyinv(x)

Equivalent to ~x in Python.

PythonCall.pyadd — Function

pyadd(x, y)

Equivalent to x + y in Python.

PythonCall.pysub — Function

pysub(x, y)

Equivalent to x - y in Python.

PythonCall.pymul — Function

pymul(x, y)

Equivalent to x * y in Python.

PythonCall.pymatmul — Function

pymatmul(x, y)

Equivalent to x @ y in Python.

PythonCall.pypow — Function

pypow(x, y, z=None)

Equivalent to x ** y or pow(x, y, z) in Python.

PythonCall.pyfloordiv — Function

pyfloordiv(x, y)

Equivalent to x // y in Python.

PythonCall.pytruediv — Function

pytruediv(x, y)

Equivalent to x / y in Python.

PythonCall.pymod — Function

pymod(x, y)

Equivalent to x % y in Python.

PythonCall.pydivmod — Function

pydivmod(x, y)

Equivalent to divmod(x, y) in Python.

PythonCall.pylshift — Function

pylshift(x, y)

Equivalent to x << y in Python.

PythonCall.pyrshift — Function

pyrshift(x, y)

Equivalent to x >> y in Python.

PythonCall.pyand — Function

pyand(x, y)

Equivalent to x & y in Python.

PythonCall.pyxor — Function

pyxor(x, y)

Equivalent to x ^ y in Python.

PythonCall.pyor — Function

pyor(x, y)

Equivalent to x | y in Python.

PythonCall.pyiadd — Function

pyiadd(x, y)

In-place add. x = pyiadd(x, y) is equivalent to x += y in Python.

PythonCall.pyisub — Function

pyisub(x, y)

In-place subtract. x = pyisub(x, y) is equivalent to x -= y in Python.

PythonCall.pyimul — Function

pyimul(x, y)

In-place multiply. x = pyimul(x, y) is equivalent to x *= y in Python.

PythonCall.pyimatmul — Function

pyimatmul(x, y)

In-place matrix multiply. x = pyimatmul(x, y) is equivalent to x @= y in Python.

PythonCall.pyipow — Function

pyipow(x, y, z=None)

In-place power. x = pyipow(x, y) is equivalent to x **= y in Python.

PythonCall.pyifloordiv — Function

pyifloordiv(x, y)

In-place floor divide. x = pyifloordiv(x, y) is equivalent to x //= y in Python.

PythonCall.pyitruediv — Function

pyitruediv(x, y)

In-place true division. x = pyitruediv(x, y) is equivalent to x /= y in Python.

PythonCall.pyimod — Function

pyimod(x, y)

In-place subtraction. x = pyimod(x, y) is equivalent to x %= y in Python.

PythonCall.pyilshift — Function

pyilshift(x, y)

In-place left shift. x = pyilshift(x, y) is equivalent to x <<= y in Python.

PythonCall.pyirshift — Function

pyirshift(x, y)

In-place right shift. x = pyirshift(x, y) is equivalent to x >>= y in Python.

PythonCall.pyiand — Function

pyiand(x, y)

In-place and. x = pyiand(x, y) is equivalent to x &= y in Python.

PythonCall.pyixor — Function

pyixor(x, y)

In-place xor. x = pyixor(x, y) is equivalent to x ^= y in Python.

PythonCall.pyior — Function

pyior(x, y)

In-place or. x = pyior(x, y) is equivalent to x |= y in Python.

Comparisons

PythonCall.pyeq — Function

pyeq(x, y)
pyeq(Bool, x, y)

Equivalent to x == y in Python. The second form converts to Bool.

PythonCall.pyne — Function

pyne(x, y)
pyne(Bool, x, y)

Equivalent to x != y in Python. The second form converts to Bool.

PythonCall.pyle — Function

pyle(x, y)
pyle(Bool, x, y)

Equivalent to x <= y in Python. The second form converts to Bool.

PythonCall.pylt — Function

pylt(x, y)
pylt(Bool, x, y)

Equivalent to x < y in Python. The second form converts to Bool.

PythonCall.pyge — Function

pyge(x, y)
pyge(Bool, x, y)

Equivalent to x >= y in Python. The second form converts to Bool.

PythonCall.pygt — Function

pygt(x, y)
pygt(Bool, x, y)

Equivalent to x > y in Python. The second form converts to Bool.

Wrapper types

PythonCall.PyList — Type

PyList{T=Py}([x])

Wraps the Python list x (or anything satisfying the sequence interface) as an AbstractVector{T}.

If x is not a Python object, it is converted to one using pylist.

PythonCall.PySet — Type

PySet{T=Py}([x])

Wraps the Python set x (or anything satisfying the set interface) as an AbstractSet{T}.

If x is not a Python object, it is converted to one using pyset.

PythonCall.PyDict — Type

PyDict{K=Py,V=Py}([x])

Wraps the Python dict x (or anything satisfying the mapping interface) as an AbstractDict{K,V}.

If x is not a Python object, it is converted to one using pydict.

PythonCall.PyIterable — Type

PyIterable{T=Py}(x)

This object iterates over iterable Python object x, yielding values of type T.

PythonCall.PyArray — Type

PyArray{T,N,M,L,R}(x; copy=true, array=true, buffer=true)

Wrap the Python array x as a Julia AbstractArray{T,N}.

The input x can be bytes, bytearray, array.array, numpy.ndarray or anything satisfying the buffer protocol (if buffer=true) or the numpy array interface (if array=true).

If copy=false then the resulting array is guaranteed to directly wrap the data in x. If copy=true then a copy is taken if necessary to produce an array.

The type parameters are all optional, and are:

T: The element type.

N: The number of dimensions.

M: True if the array is mutable.

L: True if the array supports fast linear indexing.

R: The element type of the underlying buffer. Equal to T for scalar numeric types.

PythonCall.PyIO — Type

PyIO(x; own=false, text=missing, buflen=4096)

Wrap the Python IO stream x as a Julia IO stream.

When this goes out of scope and is finalized, it is automatically flushed. If own=true then it is also closed.

If text=false then x must be a binary stream and arbitrary binary I/O is possible. If text=true then x must be a text stream and only UTF-8 must be written (i.e. use print not write). If text is not specified then it is chosen automatically. If x is a text stream and you really need a binary stream, then often PyIO(x.buffer) will work.

For efficiency, reads and writes are buffered before being sent to x. The size of the buffers is buflen. The buffers are cleared using flush.

PythonCall.PyTable — Type

PyTable

Abstract type for Python wrappers which may be interpretable as Tables.jl-compatible tables.

If pyconvert(PyTable, x::Py) is a table, then x is also a table.

PythonCall.PyPandasDataFrame — Type

PyPandasDataFrame(x; indexname=nothing, columntypes=Dict())

Wraps the pandas DataFrame x as a Tables.jl-compatible table.

indexname is the name of the column to contain the index. It may be nothing to exclude the index.

columntypes is a mapping of column names to column element types, in case automatic deduction does not work.

PythonCall.PyObjectArray — Type

PyObjectArray(undef, dims...)
PyObjectArray(array)

An array of Pys which supports the Python buffer protocol.

Internally, the objects are stored as an array of pointers.